task2
This project is a versatile blog website designed for both reading and writing posts. Users can explore a diverse array of articles on various topics and engage with content that interests them. Additionally, the platform enables users to contribute their own articles by creating and publishing posts. All content is securely stored in the backend, ensuring seamless access and management of posts. The website features a straightforward and user-friendly interface, enhancing the overall reading and writing experience for visitors.